Fix: Move heartbeatTask into separate file.
