+++ sw-mcu/canfd-to-usb-nucleo/Core/Inc/heartbeatTask.h
@@ -1,0 +1,15 @@
+/*
+ * heartbeatTask.h
+ *
+ *  Created on: Mar 31, 2020
+ *      Author: s1ckret
+ */
+
+#ifndef TASKS_INC_HEARTBEATTASK_H_
+#define TASKS_INC_HEARTBEATTASK_H_
+
+#include "cmsis_os.h"
+
+osThreadId_t * createHeartbeatTask();
+
+#endif /* TASKS_INC_HEARTBEATTASK_H_ */
